Deprecate account cli cmd
