--- packages/cli/src/cmds/account/cmds/wallet/options.ts
@@ -1,14 +1,0 @@
-import {defaultAccountPaths} from "../../paths.js";
-import {ICliCommandOptions} from "../../../../util/index.js";
-
-export interface IAccountWalletArgs {
-  walletsDir?: string;
-}
-
-export const accountWalletsOptions: ICliCommandOptions<IAccountWalletArgs> = {
-  walletsDir: {
-    description: "Directory for storing wallets.",
-    defaultDescription: defaultAccountPaths.walletsDir,
-    type: "string",
-  },
-};
