+++ packages/cli/src/cmds/validator/index.ts
@@ -1,6 +1,9 @@
 import {ICliCommand} from "../../util/index.js";
 import {IGlobalArgs} from "../../options/index.js";
-import {getAccountPaths} from "../account/paths.js";
+import {getAccountPaths} from "./paths.js";
+import {slashingProtection} from "./slashingProtection/index.js";
+import {voluntaryExit} from "./voluntaryExit.js";
+import {importCmd} from "./import.js";
 import {validatorOptions, IValidatorCliArgs} from "./options.js";
 import {validatorHandler} from "./handler.js";
 
@@ -17,4 +20,5 @@ export const validator: ICliCommand<IValidatorCliArgs, IGlobalArgs> = {
   ],
   options: validatorOptions,
   handler: validatorHandler,
+  subcommands: [slashingProtection, importCmd, voluntaryExit],
 };
