+++ packages/cli/src/cmds/validator/slashingProtection/index.ts
@@ -1,10 +1,10 @@
-import {ICliCommand} from "../../../../../util/index.js";
-import {IAccountValidatorArgs} from "../options.js";
+import {ICliCommand} from "../../../util/index.js";
+import {AccountValidatorArgs} from "../options.js";
 import {ISlashingProtectionArgs, slashingProtectionOptions} from "./options.js";
 import {importCmd} from "./import.js";
 import {exportCmd} from "./export.js";
 
-export const slashingProtection: ICliCommand<ISlashingProtectionArgs, IAccountValidatorArgs> = {
+export const slashingProtection: ICliCommand<ISlashingProtectionArgs, AccountValidatorArgs> = {
   command: "slashing-protection <command>",
   describe: "Import or export slashing protection data to or from another client.",
   options: slashingProtectionOptions,
