+++ packages/cli/src/cmds/validator/slashingProtection/options.ts
@@ -1,5 +1,5 @@
-import {ICliCommandOptions} from "../../../../../util/index.js";
-import {IValidatorCliArgs, validatorOptions} from "../../../../validator/options.js";
+import {ICliCommandOptions} from "../../../util/index.js";
+import {IValidatorCliArgs, validatorOptions} from "../options.js";
 
 export type ISlashingProtectionArgs = Pick<IValidatorCliArgs, "server"> & {
   force?: boolean;
